docker/model/python/python_bench/sql_injection_psycopg2_cursor_execute/user_activity_logger_repo/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user_activity_logger"
version = "0.1.0"
description = "用户活动日志查询服务"
dependencies = [
    "psycopg2-binary>=2.9.0",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "--junitxml=resources/test_results/results.xml"